@@ -1,5 +1,4 @@
 import 'dart:convert';
-
 import 'package:http/http.dart' as http;
 import 'package:rocket_chat_connector_flutter/exceptions/exception.dart';
 import 'package:rocket_chat_connector_flutter/models/authentication.dart';
@@ -13,6 +12,7 @@ import 'package:rocket_chat_connector_flutter/models/room.dart';
 import 'package:rocket_chat_connector_flutter/models/room_counters.dart';
 import 'package:rocket_chat_connector_flutter/models/room_messages.dart';
 import 'package:rocket_chat_connector_flutter/services/http_service.dart';
+import 'package:http_parser/http_parser.dart';
 
 class RoomService {
   HttpService _httpService;
@@ -110,5 +110,64 @@ class RoomService {
       }
     }
     throw RocketChatException(response.body);
+  }
+
+  /// Upload File to a Room
+  Future<String> uploadFile(
+    Room room,
+    String filename,
+    Authentication authentication, {
+    // A message text
+    String? msg,
+    // A description of the file
+    String? description,
+    // The thread message id (if you want upload a file to a thread)
+    String? tMid,
+    MediaType? mediaType,
+  }) async {
+    Map<String, String> fields = {};
+
+    if (msg != null) {
+      fields['msg'] = msg;
+    }
+    if (description != null) {
+      fields['description'] = description;
+    }
+    if (tMid != null) {
+      fields['tMid'] = tMid;
+    }
+    if (mediaType == null) {
+      var name =
+          filename.substring(filename.lastIndexOf('/') + 1, filename.length);
+      // 获取文件扩展名
+      List<String> fileNameSegments = name.split('.');
+      String fileExt = fileNameSegments.last.toLowerCase();
+      // 手动指定上传文件的contentType
+      if (fileExt == 'gif' ||
+          fileExt == 'jpg' ||
+          fileExt == 'jpeg' ||
+          fileExt == 'bmp' ||
+          fileExt == 'png') {
+        mediaType = MediaType('image', fileExt);
+      } else if (fileExt == 'mp4') {
+        mediaType = MediaType("video", fileExt);
+      } else {
+        mediaType = MediaType("application", "octet-stream");
+      }
+    }
+    http.StreamedResponse response = await _httpService.postFile(
+      '/api/v1/rooms.upload/${room.id}',
+      filename,
+      authentication,
+      fields: fields,
+      mediaType: mediaType,
+    );
+
+    if (response.statusCode == 200) {
+      String responseBody = await response.stream.bytesToString();
+      print(responseBody);
+      return responseBody;
+    }
+    throw RocketChatException('${response.statusCode}');
   }
 }
